Pre/postfix operator tests passing. Slight refactor removing erroneous files.
