--- transpiler/parser/test_cases_parser_input.js
@@ -1,886 +1,0 @@
-var test_cases = [
-
-
-  // ***** TEST 1 *****
-  // Swift input: 'var a = 3'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "3" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // ***** TEST 2 *****
-  // Swift input: 'var b = "hello"'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "b" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "STRING",               value: "hello" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // ***** TEST 3 *****
-  // Swift input: 'var c = true'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "c" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "BOOLEAN",              value: "true" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // ***** TEST 4 *****
-  // Swift input: 'var d = "Test this"'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "d" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "STRING",               value: "Test this" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // ***** TEST 5 *****
-  // Swift input: 'var e = ["Eggs", "Milk", "Bacon"]'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "e" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "ARRAY_START",          value: "[" },
-    { type: "STRING",               value: "Eggs" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "Milk" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "Bacon" },
-    { type: "ARRAY_END",            value: "]" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 6 *****
-  // Swift input: 'var f = ["one": 1, "two": 2, "three": 3]'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "f" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "STRING",               value: "one" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "two" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "three" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "NUMBER",               value: "3" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 7 *****
-  // Swift input: 'let g = [1 : "one",2   :"two", 3: "three"]'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "let" },
-    { type: "IDENTIFIER",           value: "g" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "STRING",               value: "one" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "STRING",               value: "two" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "NUMBER",               value: "3" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "STRING",               value: "three" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 8 *****
-  // Swift input: 'let h = 3.14'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "let" },
-    { type: "IDENTIFIER",           value: "h" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "3.14" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 9 *****
-  // Swift input: 'let i = 5+6'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "let" },
-    { type: "IDENTIFIER",           value: "i" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "5" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "NUMBER",               value: "6" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 10 *****
-  // Swift input: 'var j = 5 + 6 / 4 - (-16 % 4.2) * 55'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "j" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "5" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "NUMBER",               value: "6" },
-    { type: "OPERATOR",             value: "/" },
-    { type: "NUMBER",               value: "4" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "PUNCTUATION",          value: "(" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "NUMBER",               value: "16" },
-    { type: "OPERATOR",             value: "%" },
-    { type: "NUMBER",               value: "4.2" },
-    { type: "PUNCTUATION",          value: ")" },
-    { type: "OPERATOR",             value: "*" },
-    { type: "NUMBER",               value: "55" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 11 *****
-  // Swift input: 'var k = "Stephen" + " " + "Tabor" + "!"'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "k" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "STRING",               value: "Stephen" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "STRING",               value: " " },
-    { type: "OPERATOR",             value: "+" },
-    { type: "STRING",               value: "Tabor" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "STRING",               value: "!" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 12 *****
-  // Swift input: 'let l = 6 !== 9'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "let" },
-    { type: "IDENTIFIER",           value: "l" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "6" },
-    { type: "OPERATOR",             value: "!" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "9" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 13 *****
-  // Swift input: 'var a = 1; var m = ++a;'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "m" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 14 *****
-  // Swift input: 'var a = 1; var n = a++;'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "n" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 15 *****
-  // Swift input: 'var planet = "Earth"; let o = "Hello \\(planet)!"'
-  [
-    { type: "DECLARATION_KEYWORD",        value: "var" },
-    { type: "IDENTIFIER",                 value: "planet" },
-    { type: "OPERATOR",                   value: "=" },
-    { type: "STRING",                     value: "Earth" },
-    { type: "PUNCTUATION",                value: ";" },
-    { type: "DECLARATION_KEYWORD",        value: "let" },
-    { type: "IDENTIFIER",                 value: "o" },
-    { type: "OPERATOR",                   value: "=" },
-    { type: "STRING",                     value: "Hello " },
-    { type: "STRING_INTERPOLATION_START", value: "\\(" },
-    { type: "IDENTIFIER",                 value: "planet" },
-    { type: "STRING_INTERPOLATION_END",   value: ")" },
-    { type: "STRING",                     value: "!" },
-    { type: "TERMINATOR",                 value: "EOF" }
-  ],
-
-  // ***** TEST 16 *****
-  // Swift input: 'var planet = "Earth"; let o = "\\(planet)"'
-  [
-    { type: "DECLARATION_KEYWORD",        value: "var" },
-    { type: "IDENTIFIER",                 value: "planet" },
-    { type: "OPERATOR",                   value: "=" },
-    { type: "STRING",                     value: "Earth" },
-    { type: "PUNCTUATION",                value: ";" },
-    { type: "DECLARATION_KEYWORD",        value: "let" },
-    { type: "IDENTIFIER",                 value: "o" },
-    { type: "OPERATOR",                   value: "=" },
-    { type: "STRING",                     value: "" },
-    { type: "STRING_INTERPOLATION_START", value: "\\(" },
-    { type: "IDENTIFIER",                 value: "planet" },
-    { type: "STRING_INTERPOLATION_END",   value: ")" },
-    { type: "STRING",                     value: "" },
-    { type: "TERMINATOR",                 value: "EOF" }
-  ],
-
-
-  // ***** TEST 17 *****
-  // Swift input: 'var p = "\\(100 - 99), 2, 3"'
-  [
-    { type: "DECLARATION_KEYWORD",        value: "var" },
-    { type: "IDENTIFIER",                 value: "p" },
-    { type: "OPERATOR",                   value: "=" },
-    { type: "STRING",                     value: "" },
-    { type: "STRING_INTERPOLATION_START", value: "\\(" },
-    { type: "NUMBER",                     value: "100" },
-    { type: "OPERATOR",                   value: "-" },
-    { type: "NUMBER",                     value: "99" },
-    { type: "STRING_INTERPOLATION_END",   value: ")" },
-    { type: "STRING",                     value: ", 2, 3" },
-    { type: "TERMINATOR",                 value: "EOF" }
-  ],
-
-  // ***** TEST 18 *****
-  // Swift input: 'let q = ["array1": [1,2,3], "array2": [4,5,6]];'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "let" },
-    { type: "IDENTIFIER",           value: "q" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "STRING",               value: "array1" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "ARRAY_START",          value: "[" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "NUMBER",               value: "3" },
-    { type: "ARRAY_END",            value: "]" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "array2" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "ARRAY_START",          value: "[" },
-    { type: "NUMBER",               value: "4" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "NUMBER",               value: "5" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "NUMBER",               value: "6" },
-    { type: "ARRAY_END",            value: "]" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 19 *****
-  // Swift input: 'var s = arr[0];'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },//
-    { type: "IDENTIFIER",           value: "arr" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "ARRAY_START",          value: "[" },
-    { type: "STRING",               value: "Eggs" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "Milk" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "Bacon" },
-    { type: "ARRAY_END",            value: "]" },
-    { type: "PUNCTUATION",             value: ";" },//
-
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "s" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "IDENTIFIER",           value: "arr" },
-    { type: "PUNCTUATION",          value: "[" },
-    { type: "NUMBER",               value: "0" },
-    { type: "PUNCTUATION",          value: "]" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 20 *****
-  // Swift input: 'let arr = [1, 2]; let t = 100; var u = arr[t - 99];'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },//
-    { type: "IDENTIFIER",           value: "arr" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "ARRAY_START",          value: "[" },
-    { type: "STRING",               value: "Eggs" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "Milk" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "STRING",               value: "Bacon" },
-    { type: "ARRAY_END",            value: "]" },
-    { type: "PUNCTUATION",             value: ";" },//
-
-    { type: "DECLARATION_KEYWORD",  value: "let" },
-    { type: "IDENTIFIER",           value: "t" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "100" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "u" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "IDENTIFIER",           value: "arr" },
-    { type: "PUNCTUATION",          value: "[" },
-    { type: "IDENTIFIER",           value: "t" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "NUMBER",               value: "99" },
-    { type: "PUNCTUATION",          value: "]" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 21 ***** TODO Did not pass
-  // Swift input: 'let arr = [1,2]; var v = [ arr[0]: [[1,2], [3,4]], arr[1]: [["one", "two"], ["three", "four"]] ];'
-  // Swift input: let arr = [1,2]; var v = [  ];
-  // Swift input: v[arr[0]] = [[1,2], [3,4]];
-  // Swift input: v[arr[1]] = [["one", "two"], ["three", "four"]];
-  [
-    // NEW CODE That needs to be transferred from lower token set to this one in CleanUp function
-     { type: "DECLARATION_KEYWORD",  value: "let" },
-     { type: "IDENTIFIER",           value: "arr" },
-     { type: "OPERATOR",             value: "=" },
-     { type: "ARRAY_START",          value: "[" },
-     { type: "NUMBER",               value: "1" },
-     { type: "PUNCTUATION",          value: "," },
-     { type: "NUMBER",               value: "2" },
-     { type: "ARRAY_END",            value: "]" },
-     { type: "PUNCTUATION",          value: ";" },
-     { type: "DECLARATION_KEYWORD",  value: "var" },
-     { type: "IDENTIFIER",           value: "v" },
-     { type: "OPERATOR",             value: "=" },
-     { type: "DICTIONARY_START",     value: "[" },
-     { type: "DICTIONARY_END",       value: "]" },
-     { type: "PUNCTUATION",          value: ";" },
-     { type: "IDENTIFIER",           value: "v" },
-     { type: "PUNCTUATION",          value: "[" },
-     { type: "IDENTIFIER",           value: "arr" },
-     { type: "PUNCTUATION",          value: "[" },
-     { type: "NUMBER",               value: "0" },
-     { type: "PUNCTUATION",          value: "]" },
-     { type: "PUNCTUATION",          value: "]" },
-     { type: "OPERATOR",             value: "=" },
-     { type: "ARRAY_START",          value: "[" },
-     { type: "ARRAY_START",          value: "[" },
-     { type: "NUMBER",               value: "1" },
-     { type: "PUNCTUATION",          value: "," },
-     { type: "NUMBER",               value: "2" },
-     { type: "ARRAY_END",            value: "]" },
-     { type: "PUNCTUATION",          value: "," },
-     { type: "ARRAY_START",          value: "[" },
-     { type: "NUMBER",               value: "3" },
-     { type: "PUNCTUATION",          value: "," },
-     { type: "NUMBER",               value: "4" },
-     { type: "ARRAY_END",            value: "]" },
-     { type: "ARRAY_END",            value: "]" },
-     { type: "PUNCTUATION",          value: ";" },
-
-     { type: "IDENTIFIER",           value: "v" },
-     { type: "PUNCTUATION",          value: "[" },
-     { type: "IDENTIFIER",           value: "arr" },
-     { type: "PUNCTUATION",          value: "[" },
-     { type: "NUMBER",               value: "1" },
-     { type: "PUNCTUATION",          value: "]" },
-     { type: "PUNCTUATION",          value: "]" },
-     { type: "OPERATOR",             value: "=" },
-     { type: "ARRAY_START",          value: "[" },
-     { type: "ARRAY_START",          value: "[" },
-     { type: "STRING",               value: "one" },
-     { type: "PUNCTUATION",          value: "," },
-     { type: "STRING",               value: "two" },
-     { type: "ARRAY_END",            value: "]" },
-     { type: "PUNCTUATION",          value: "," },
-     { type: "ARRAY_START",          value: "[" },
-     { type: "STRING",               value: "three" },
-     { type: "PUNCTUATION",          value: "," },
-     { type: "STRING",               value: "four" },
-     { type: "ARRAY_END",            value: "]" },
-     { type: "ARRAY_END",            value: "]" },
-     { type: "PUNCTUATION",          value: ";" },
-     { type: "TERMINATOR",           value: "EOF" }
-
-    //{ type: "DECLARATION_KEYWORD",  value: "let" },
-    //{ type: "IDENTIFIER",           value: "arr" },
-    //{ type: "OPERATOR",             value: "=" },
-    //{ type: "ARRAY_START",          value: "[" },
-    //{ type: "NUMBER",               value: "1" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "NUMBER",               value: "2" },
-    //{ type: "ARRAY_END",            value: "]" },
-    //{ type: "PUNCTUATION",          value: ";" },
-    //{ type: "DECLARATION_KEYWORD",  value: "var" },
-    //{ type: "IDENTIFIER",           value: "v" },
-    //{ type: "OPERATOR",             value: "=" },
-    //{ type: "DICTIONARY_START",     value: "[" },
-    //{ type: "IDENTIFIER",           value: "arr" },
-    //{ type: "PUNCTUATION",          value: "[" },
-    //{ type: "NUMBER",               value: "0" },
-    //{ type: "PUNCTUATION",          value: "]" },
-    //{ type: "PUNCTUATION",          value: ":" },
-    //{ type: "ARRAY_START",          value: "[" },
-    //{ type: "ARRAY_START",          value: "[" },
-    //{ type: "NUMBER",               value: "1" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "NUMBER",               value: "2" },
-    //{ type: "ARRAY_END",            value: "]" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "ARRAY_START",          value: "[" },
-    //{ type: "NUMBER",               value: "3" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "NUMBER",               value: "4" },
-    //{ type: "ARRAY_END",            value: "]" },
-    //{ type: "ARRAY_END",            value: "]" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "IDENTIFIER",           value: "arr" },
-    //{ type: "PUNCTUATION",          value: "[" },
-    //{ type: "NUMBER",               value: "1" },
-    //{ type: "PUNCTUATION",          value: "]" },
-    //{ type: "PUNCTUATION",          value: ":" },
-    //{ type: "ARRAY_START",          value: "[" },
-    //{ type: "ARRAY_START",          value: "[" },
-    //{ type: "STRING",               value: "one" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "STRING",               value: "two" },
-    //{ type: "ARRAY_END",            value: "]" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "ARRAY_START",          value: "[" },
-    //{ type: "STRING",               value: "three" },
-    //{ type: "PUNCTUATION",          value: "," },
-    //{ type: "STRING",               value: "four" },
-    //{ type: "ARRAY_END",            value: "]" },
-    //{ type: "ARRAY_END",            value: "]" },
-    //{ type: "DICTIONARY_END",       value: "]" },
-    //{ type: "PUNCTUATION",          value: ";" },
-    //{ type: "TERMINATOR",           value: "EOF" }
-  ],
-
-  // ***** TEST 22 *****
-  // Swift input: 'var w = [1: [[1: "two"], [3: "four"]], 2: [["one": 2], ["three": 4]]];'
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "w" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "ARRAY_START",          value: "[" },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "STRING",               value: "two" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "NUMBER",               value: "3" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "STRING",               value: "four" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "ARRAY_END",            value: "]" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "ARRAY_START",          value: "[" },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "STRING",               value: "one" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "NUMBER",               value: "2" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "PUNCTUATION",          value: "," },
-    { type: "DICTIONARY_START",     value: "[" },
-    { type: "STRING",               value: "three" },
-    { type: "PUNCTUATION",          value: ":" },
-    { type: "NUMBER",               value: "4" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "ARRAY_END",            value: "]" },
-    { type: "DICTIONARY_END",       value: "]" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF" }
-  ],
-
-//   [
-//       { type: "DECLARATION_KEYWORD",  value: "let" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "NUMBER",               value: "1" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "NUMBER",               value: "2" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "DECLARATION_KEYWORD",  value: "var" },
-//       { type: "IDENTIFIER",           value: "s" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "SUBSTRING_LOOKUP",     value: "[" },
-//       { type: "NUMBER",               value: "0" },
-//       { type: "SUBSTRING_LOOKUP",     value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "TERMINATOR",           value: "EOF" }
-//     ],
-//     // 20
-//     [
-//       { type: "DECLARATION_KEYWORD",  value: "let" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "NUMBER",               value: "1" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "NUMBER",               value: "2" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "DECLARATION_KEYWORD",  value: "let" },
-//       { type: "IDENTIFIER",           value: "t" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "NUMBER",               value: "100" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "DECLARATION_KEYWORD",  value: "var" },
-//       { type: "IDENTIFIER",           value: "u" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "SUBSTRING_LOOKUP",     value: "[" },
-//       { type: "IDENTIFIER",           value: "t" },
-//       { type: "OPERATOR",             value: "-" },
-//       { type: "NUMBER",               value: "99" },
-//       { type: "SUBSTRING_LOOKUP",     value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "TERMINATOR",           value: "EOF" }
-//     ],
-//     // 21
-//     [
-//       { type: "DECLARATION_KEYWORD",  value: "let" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "NUMBER",               value: "1" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "NUMBER",               value: "2" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "DECLARATION_KEYWORD",  value: "var" },
-//       { type: "IDENTIFIER",           value: "u" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "SUBSTRING_LOOKUP",     value: "[" },
-//       { type: "NUMBER",               value: "0" },
-//       { type: "SUBSTRING_LOOKUP",     value: "]" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "TERMINATOR",           value: "EOF" }
-//     ],
-//     // 22
-//     [
-//       { type: "DECLARATION_KEYWORD",  value: "let" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "NUMBER",               value: "1" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "NUMBER",               value: "2" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "DECLARATION_KEYWORD",  value: "var" },
-//       { type: "IDENTIFIER",           value: "v" },
-//       { type: "OPERATOR",             value: "=" },
-//       { type: "DICTIONARY_START",     value: "[" },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "SUBSTRING_LOOKUP",     value: "[" },
-//       { type: "NUMBER",               value: "0" },
-//       { type: "SUBSTRING_LOOKUP",     value: "]" },
-//       { type: "PUNCTUATION",          value: ":" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "NUMBER",               value: "1" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "NUMBER",               value: "2" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "NUMBER",               value: "3" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "NUMBER",               value: "4" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "IDENTIFIER",           value: "arr" },
-//       { type: "SUBSTRING_LOOKUP",     value: "[" },
-//       { type: "NUMBER",               value: "1" },
-//       { type: "SUBSTRING_LOOKUP",     value: "]" },
-//       { type: "PUNCTUATION",          value: ":" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "STRING",               value: "one" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "STRING",               value: "two" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "ARRAY_START",          value: "[" },
-//       { type: "STRING",               value: "three" },
-//       { type: "PUNCTUATION",          value: "," },
-//       { type: "STRING",               value: "four" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "ARRAY_END",            value: "]" },
-//       { type: "DICTIONARY_END",       value: "]" },
-//       { type: "PUNCTUATION",          value: ";" },
-//       { type: "TERMINATOR",           value: "EOF" }
-//     ]
-//   ]
-// };
-
-
-
-  // ***** TEST 23 *****
-  // should handle single-line if statements'
-  // Swift input: 'var a = 5; if (true) {--a};';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "5" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "PUNCTUATION",          value: "(" },
-    { type: "BOOLEAN",              value: "true" },
-    { type: "PUNCTUATION",          value: ")" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-
-  // ***** TEST 24 *****
-  // should handle single-line if statements with multi-character logical operators
-  // Swift input: 'var b = 6; if (5 <= 6) {b++};';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "b" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "6" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "PUNCTUATION",          value: "(" },
-    { type: "NUMBER",               value: "5" },
-    { type: "OPERATOR",             value: "<" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "6" },
-    { type: "PUNCTUATION",          value: ")" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "b" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // ***** TEST 25 *****
-  // should handle single-line if statements with multi-character logical operators and multi-character mathematical operators
-  // Swift input: 'var c = 1; if (c == 1) {c *= 5};';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "c" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "PUNCTUATION",          value: "(" },
-    { type: "IDENTIFIER",           value: "c" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ")" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "c" },
-    { type: "OPERATOR",             value: "*" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "5" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // ***** TEST 26 *****
-  // should handle single-line if statements without a parenthetical
-  // Swift input: 'var d = 1; if d != 2 {d++};';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "d" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "IDENTIFIER",           value: "d" },
-    { type: "OPERATOR",             value: "!" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "d" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // ***** TEST 27 *****
-  // should handle complex conditionals without an outer parenthetical
-  // Swift input: 'var e = 1; if (e + 1) == 2 {e = 5};';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "e" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "PUNCTUATION",          value: "(" },
-    { type: "IDENTIFIER",           value: "e" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ")" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "e" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "5" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  //it('should handle single line if/else statements', function() {
-  //  input = 'var f = true; if !f {f = true} else {f = false};';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "f" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "BOOLEAN",              value: "true" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "OPERATOR",             value: "!" },
-    { type: "IDENTIFIER",           value: "f" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "f" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "BOOLEAN",              value: "true" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "STATEMENT_KEYWORD",    value: "else" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "f" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "BOOLEAN",              value: "false" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  // it('should handle single-line if/else-if/else statements with parentheticals', function() {
-  //  input = 'var a = 1; if (1 > 2) {++a} else if (1 < 2) {--a} else {a = 42}';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "PUNCTUATION",          value: "(" },
-    { type: "NUMBER",               value: "1" },
-    { type: "OPERATOR",             value: ">" },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: ")" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "STATEMENT_KEYWORD",    value: "else" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "PUNCTUATION",          value: "(" },
-    { type: "NUMBER",               value: "1" },
-    { type: "OPERATOR",             value: "<" },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: ")" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "STATEMENT_KEYWORD",    value: "else" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "42" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ],
-
-  //input = 'var a = 1; if 1 > 2 {++a} else if 1 < 2 {--a} else {a = 42}';
-  [
-    { type: "DECLARATION_KEYWORD",  value: "var" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "1" },
-    { type: "PUNCTUATION",          value: ";" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "NUMBER",               value: "1" },
-    { type: "OPERATOR",             value: ">" },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "OPERATOR",             value: "+" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "STATEMENT_KEYWORD",    value: "else" },
-    { type: "STATEMENT_KEYWORD",    value: "if" },
-    { type: "NUMBER",               value: "1" },
-    { type: "OPERATOR",             value: "<" },
-    { type: "NUMBER",               value: "2" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "OPERATOR",             value: "-" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "STATEMENT_KEYWORD",    value: "else" },
-    { type: "PUNCTUATION",          value: "{" },
-    { type: "IDENTIFIER",           value: "a" },
-    { type: "OPERATOR",             value: "=" },
-    { type: "NUMBER",               value: "42" },
-    { type: "PUNCTUATION",          value: "}" },
-    { type: "TERMINATOR",           value: "EOF"}
-  ]
-
-];
-
-module.exports = test_cases;
